+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - 2017-MM-DD
+### Added
+- Dry run option to db.initializeDatabase
+### Removed
+- Removed SQLite support
 
 ## [0.2.1] - 2017-09-03
 ### Fixed